Add edit button to tasks in to-do list and fix bug
Edit button will now present a form to edit the tasks in the list.
Upon refresh, when a task would be added using the add task button, it would remove/mutate the previous tasks in local memory. Now it works as intended and adds new tasks on top of current tasks in memory.